refactor: introduce a new common module for shared utils to avoid cyclic dependencies and code duplication
