+++ build.gradle
@@ -35,12 +35,22 @@ jacoco {
     toolVersion = "0.7.6.201602180812"
 }
 
-repositories {
-    mavenCentral()
+
+allprojects{
+    sourceCompatibility = jvmTargetVersion
+    targetCompatibility = jvmTargetVersion
+    repositories {
+        mavenCentral()
+    }
+    tasks.withType(JavaCompile) {
+        options.encoding = 'UTF-8'
+        if (!javaHome) {
+            javaHome =  System.env.JAVA_HOME
+        }
+        options.bootClasspath = "$javaHome/jre/lib/rt.jar"
+    }
 }
 
-sourceCompatibility = jvmTargetVersion
-targetCompatibility = jvmTargetVersion
 
 testSets {
     integrationTest
@@ -109,6 +119,12 @@ jacocoTestReport {
 }
 
 dependencies {
+    compile(project(':testme-intellij-common')){
+        exclude group:'com.jetbrains', module: 'ideaIC'
+        exclude group:'org.jetbrains.plugins', module: 'junit'
+        exclude group:'org.jetbrains.plugins', module: 'properties'
+        exclude group:'org.jetbrains.plugins', module: 'Groovy'
+    }
     compile(project(':testme-intellij-groovy')){
         exclude group:'com.jetbrains', module: 'ideaIC'
         exclude group:'org.jetbrains.plugins', module: 'junit'
@@ -131,13 +147,6 @@ dependencies {
 //subprojects{
 //    ideaVersion=ideaVersion
 //}
-tasks.withType(JavaCompile) {
-    options.encoding = 'UTF-8'
-    if (!javaHome) {
-        javaHome =  System.env.JAVA_HOME
-    }
-    options.bootClasspath = "$javaHome/jre/lib/rt.jar"
-}
 
 tasks.withType(Test) {
     systemProperty 'enableIdeaGroovyPlugin', enableIdeaGroovyPlugin
