+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'testme-intellij'
-include 'testme-intellij-groovy', 'testme-intellij-scala'
+include 'testme-intellij-common', 'testme-intellij-groovy', 'testme-intellij-scala'
