+++ testme-intellij-common/build.gradle
@@ -1,0 +1,25 @@
+
+apply plugin: 'java'
+apply plugin: 'org.jetbrains.intellij'
+
+
+def ideaVersion = (rootProject.hasProperty('ideaVersion') && rootProject.ideaVersion) ? rootProject.ideaVersion : '2016.3'
+def ideaType = (rootProject.hasProperty('ideaType') && rootProject.ideaType) ? rootProject.ideaType : 'IC'
+def spockVersion = ideaVersion.startsWith("14.") || ideaVersion.startsWith("15.")? "1.0-groovy-2.3" : "1.0-groovy-2.4"
+
+intellij {
+    pluginName 'TestMe'
+    instrumentCode = true
+    version ideaVersion
+    type ideaType
+    downloadSources true
+    sameSinceUntilBuild false
+    updateSinceUntilBuild false
+    plugins = ['junit','properties']
+}
+
+dependencies {
+    testCompile("org.spockframework:spock-core:$spockVersion"){
+        exclude group: 'org.codehaus.groovy'
+    }
+}
