+++ testme-intellij-groovy/build.gradle
@@ -42,9 +42,11 @@ tasks.withType(JavaCompile) {
     options.bootClasspath = "$javaHome/jre/lib/rt.jar"
 }
 
-//dependencies {
-//    compileOnly 'org.jetbrains.plugins:Groovy:${ideaVersion}'
-//    compile 'org.jetbrains.plugins:Groovy:${ideaVersion}'
-//    compile 'com.jetbrains:ideaIC:${ideaVersion}'
-//    compile 'com.jetbrains:ideaIC:14.1.2'
-//}
+dependencies {
+    compile(project(':testme-intellij-common')){
+        exclude group:'com.jetbrains', module: 'ideaIC'
+        exclude group:'org.jetbrains.plugins', module: 'junit'
+        exclude group:'org.jetbrains.plugins', module: 'properties'
+        exclude group:'org.jetbrains.plugins', module: 'Groovy'
+    }
+}
