+++ testme-intellij-scala/build.gradle
@@ -29,6 +29,12 @@ intellij {
     plugins = ['junit','properties',"org.intellij.scala:$scalaPluginVersion"]
 }
 dependencies {
+    compile(project(':testme-intellij-common')){
+        exclude group:'com.jetbrains', module: 'ideaIC'
+        exclude group:'org.jetbrains.plugins', module: 'junit'
+        exclude group:'org.jetbrains.plugins', module: 'properties'
+        exclude group:'org.jetbrains.plugins', module: 'Groovy'
+    }
     testCompile("org.spockframework:spock-core:$spockVersion"){
         exclude group: 'org.codehaus.groovy'
     }
